pyproject.toml
[tool.poetry]
name = "stock-prediction-pipelines"
version = "0.1.0"
description = "Stock Prediction Pipelines"
authors = ["Josh Phillips <Josh.L.Phillips@me.com>"]

# Important: Enable virtual environments within the project
[virtualenvs]
in-project = true

[tool.poetry.dependencies]
python = "3.12.6"
requests = "^2.28.1"
psycopg2-binary = "^2.9.3"
boto3 = "^1.26.0"
scikit-learn = "^1.0.2"
apache-airflow = "2.10.1"
#apache-airflow-providers-openlineage = "1.11.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# Deployment Scripts